separate data from component
