+++ data/navItems.js
@@ -1,0 +1,9 @@
+const navbarItems = [
+    { id: '#about', title: 'About' },
+    { id: '#chapters', title: 'Chapters' },
+    { id: '#reviews', title: 'Reviews' },
+    { id: '#author', title: 'Author' },
+    { id: '#contact', title: 'Contact' },
+];
+
+export default navbarItems;
